# app/src/lib/api/model/createOrderDto.ts
/**
 * Generated by orval v7.15.0 🍺
 * Do not edit manually.
 * Market API
 * Документация к моему API
 * OpenAPI spec version: 1.0
 */

export interface CreateOrderDto {
  /** ID лота, который хотят купить */
  lotId: string;
}