# app/src/lib/api/model/loginDto.ts
/**
 * Generated by orval v7.15.0 🍺
 * Do not edit manually.
 * Market API
 * Документация к моему API
 * OpenAPI spec version: 1.0
 */

export interface LoginDto {
  /** Email пользователя */
  email: string;
  /** Пароль пользователя */
  password: string;
}