# app/src/lib/api/model/lotDto.ts
/**
 * Generated by orval v7.15.0 🍺
 * Do not edit manually.
 * Market API
 * Документация к моему API
 * OpenAPI spec version: 1.0
 */

export interface LotDto {
  /** Название продукта */
  title: string;
  /** Стоимость продукта */
  cost: string;
  /** Тип продукта */
  type: string;
  /** Описание продукта */
  description: string;
}