# app/src/lib/api/model/openShopDto.ts
/**
 * Generated by orval v7.15.0 🍺
 * Do not edit manually.
 * Market API
 * Документация к моему API
 * OpenAPI spec version: 1.0
 */

export interface OpenShopDto {
  /** Название магазина */
  title: string;
  /** Тип магазина */
  type: string;
  /** Описание магазина */
  description: string;
  /** Номер телефона продавца */
  phone: string;
}