# app/src/lib/api/model/requestRegisterDto.ts
/**
 * Generated by orval v7.15.0 🍺
 * Do not edit manually.
 * Market API
 * Документация к моему API
 * OpenAPI spec version: 1.0
 */

export interface RequestRegisterDto {
  /** Login пользователя */
  login: string;
  /** Email пользователя */
  email: string;
  /** Password пользователя */
  password: string;
}