# app/src/lib/api/model/resendCodeDto.ts
/**
 * Generated by orval v7.15.0 🍺
 * Do not edit manually.
 * Market API
 * Документация к моему API
 * OpenAPI spec version: 1.0
 */

export interface ResendCodeDto {
  /** ID записи о верификации (из /initiate) */
  verificationId: string;
}