# app/src/lib/api/model/reviewsDto.ts
/**
 * Generated by orval v7.15.0 🍺
 * Do not edit manually.
 * Market API
 * Документация к моему API
 * OpenAPI spec version: 1.0
 */

export interface ReviewsDto {
  /** Text-content отзыва */
  description: string;
  /** Рейтинг от 1 до 5 */
  rating: number;
  /** ID магазина (продавца), к которому отзыв */
  shopId: string;
  /** ID лота (если отзыв о лоте) */
  lotId?: string;
}