# app/src/lib/api/model/verifyRegisterDto.ts
/**
 * Generated by orval v7.15.0 🍺
 * Do not edit manually.
 * Market API
 * Документация к моему API
 * OpenAPI spec version: 1.0
 */

export interface VerifyRegisterDto {
  /** Email пользователя */
  email: string;
  /** Code подтверждения с почты */
  code: string;
}